+++ sphinx/application.py
@@ -77,6 +77,7 @@ builtin_extensions = (
     'sphinx.config',
     'sphinx.domains.c',
     'sphinx.domains.changeset',
+    'sphinx.domains.citation',
     'sphinx.domains.cpp',
     'sphinx.domains.javascript',
     'sphinx.domains.math',
